Fix: introduce `guard_failure` callback argument to `Builder.walk()`.
The problem occurs when the hot path contains two guards and the
optimizer decides to reorder them. In this case, the guard that is
delayed gets replicated on the cold paths of the guard that is advanced.
Its GuardFailure must therefore be provided to those cold paths.

This fix is necessary, but not sufficient. There is still a bug to be found.