Don't allow the optimizer to reorder guards.
This is in bad taste, in the sense that the optimizer "should" ignore
all parts of the input program that do not affect its meaning, whereas
this change explicitly preserves a feature of the input program in the
output program. However, at the moment, the optimizer is often making
quite bad decisions about the order of the guard instructions, so it's
better not to give it that freedom.