WIP: Redesign `Jit::new_entry()`.
We construct two `Case`s, and make them jump to each other. The "private" one
has the standard external `Convention`, and the "public" one has the
`Convention` requested by the caller of `new_entry()`.

These two jumps are not meant to be an infinite loop. Each time we jump to the
private `Case`, we want to run the lowerer's epilogue, set the return value,
exit Mijit, wait to for Mijit to be re-entered, and then run the lowerer's
prologue. This is all TODO.